refactor(region): extract PropositionsSyncService — #828 Step 1
First bounded-context split toward #828. Extracts the propositions sync
lifecycle (discover → extract_and_upsert → analysis + backfill +
regenerate) out of RegionSyncService into a dedicated service.

region-sync.service.ts shrinks from 3,929 → 3,538 LOC. PropositionsSyncService
is 282 LOC, owns the proposition-specific concerns end-to-end. The
orchestrator keeps the two public-ish methods as thin delegates:

  - syncPropositions: builds stage patterns (because the stage helpers
    still live in the orchestrator pending the shared-helpers extraction
    — see follow-up sub-task in #828), then forwards to
    propositionsSyncService.sync(provider, jobId, stagePatterns, upsertCb)
  - regeneratePropositionAnalysis: forwards to propositionsSyncService.regenerate(id)

The extracted service takes its full dependency graph via constructor
injection (DbService + @Optional propositionAnalysis + @Optional
cacheService). The orchestrator passes its own upsertByExternalId helper
in as a callback so behavior — including the 'propositions:' cache
prefix invalidation — is byte-identical.

Public API unchanged: RegionDomainService.regeneratePropositionAnalysis
still delegates through RegionSyncService, the resolver path
(region.resolver.ts:199) is untouched. All 2,105 backend tests pass with
no behavior changes — including the 14 propositions-flavored tests that
exercise the delegate path (region-sync.service.spec.ts test modules now
register PropositionsSyncService).

Temporary local duplications kept tight for review-ability:
  - resolveStageFromStatus (4 LOC pure function) is duplicated inline
    in propositions-sync.service.ts; consolidated when the shared stage
    helper extraction lands as a follow-up step
  - StagePattern + UpsertByExternalId types are re-declared as local
    interfaces; same follow-up will move them to a shared module

Refs #828
